Add Skylight instrumentation
I discussed with @sethherr about adding Skylight instrumentation
to the Bike Index website under the new [Skylight for Open Source](https://www.skylight.io/oss)
program.

If you’re not already familiar with [Skylight](https://www.skylight.io),
it is a smart profiler for Rails apps. Skylight makes it easy to
pinpoint performance issues in Rails applications.

We work on a lot of open source projects ourselves, and in our
experience it can be pretty hard to get contributors to work on
application performance issues. Few contributors consider working
on performance problems, and the ones that might be interested may
not even know where to start.

By making performance information more accessible, we hope to
inspire potential contributors to tackle slow parts of your app,
and have a good way to see if their contributions helped.

Once this patch is merged and deployed*, you will be able to view the
performance data we collected at the [public Skylight dashboard](https://oss.skylight.io/app/applications/j93iQ4K2pxCP).
The dashboard will be accessible to anyone (no Skylight account
required) to make it easy for contributors.

(*Actually, I lied a little. We still need to set the
`SKYLIGHT_AUTHENTICATION` environment variable to the appropriate API
key on production, but I will work with @sethherr on that off-thread.)