Externalize controller flash strings (#1319)
Externalizes flash strings from ~half our controllers

Deepens `ControllerHelpers#translation` so it infers the method invoking it rather than the controller action in which it's invoked. This is useful because a private helper method or before action can be invoked in the context of multiple controller actions but we need a 1:1 mapping from call site to translation file key.