Use render :plain instead of render :text
DEPRECATION WARNING: `render :text` is deprecated because it does not actually render a `text/plain` response.
Switch to `render plain: 'plain text'` to render as `text/plain`,
`render html: '<strong>HTML</strong>'` to render as `text/html`,
or `render body: 'raw'` to match the deprecated behavior and render with the default Content-Type,
which is `text/html`.