Specify Rails release in existing migrations
```
StandardError: An error has occurred, this and all later migrations
canceled:

Directly inheriting from ActiveRecord::Migration is not supported.
Please specify the Rails release the migration was written for:

Caused by:

StandardError: Directly inheriting from ActiveRecord::Migration is not
supported. Please specify the Rails release the migration was written
for:
```