Simplify Geocodeable, geocode by address field (#1512)
This PR is to simplify the `Geocodeable` mixin by consistently geocoding using an `address` column, adding one to geocodeable models as needed. 

For models where the `address` value is derived from other fields, we set the value in a `before_validation` `set_address` hook.

This patch renames the existing `StolenRecord#address` method to `#display_address` in order to retain its current functionality.