Add hotwire-livereload (#2759)
[hotwire-livereload](https://github.com/kirillplatonov/hotwire-livereload/)
locked to version 1.4.1 - the last version prior to it auto injecting
the code.

Specifically this is because it double renders lookbook and throws an
error, but also other places in the app where Turbo isn't enabled.

See https://github.com/kirillplatonov/hotwire-livereload/issues/73 -
hoping to figure out a way to disable for just some layouts.